Restructure
Add mode and signature fields to Scale struct and remove notes field.
Change Scale::new() function to simply return a Scale struct with these
fields populated, and use enumerate function to construct the notes of
the scale using that information.

As of this commit the only available Mode variants are Ionian (major),
Aeolin (minor), and chromatic.

Replace constants with functions implemented on Mode and Signature
enums.

Change test file to work with new Scale.mode field rather than inputting
interval string.

Remove Scale::chromatic function and enable the Scale::enumerate
function to handle the chromatic as well as diatonic case. Chromatic is
implemented as a variant of the mode enum.